Use -m68030 instead of -m68020-60
-m68020-60 basically means
"don't use 020+ instructions that are missing on 060 and
 optimize for 060 while still make the whole thing run on 020"

As an example the compiler will then prefer mulu over a few
shifts and adds, because that is faster on 060. But it'll be
slow on 020/030.

ADoom has CPU specific assembly implementations for the most
importend functions anyway, so I'd rather let the compiler
optmize the rest of the C code for the lower end machines.
Thus switch to -m68030.